backend/requirements.txt
# FastAPI and server
fastapi>=0.115.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
alembic>=1.13.1

# Auth
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0

# Validation and settings
pydantic[email]>=2.5.3
pydantic-settings>=2.1.0
email-validator>=2.1.0

# DocuSign
docusign-esign>=3.25.0

# AWS S3 (file storage)
boto3>=1.34.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.26.0
